+++ src/http/hiochainstream.h
@@ -17,7 +17,7 @@ public:
     virtual void continueRead();
     virtual void suspendRead();
     virtual int sendRespHeaders(HttpRespHeaders *pHeaders);
-    virtual int sendfile(int fdSrc, off_t off, size_t size);
+    virtual int sendfile(int fdSrc, off_t off, off_t size);
     virtual int read(char *pBuf, int size);
     virtual int close();
     virtual int flush();
